NobHUB/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
google-auth==2.40.2
google-genai==1.18.0
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
requests==2.32.3
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.39
typing-inspection==0.4.1
typing_extensions==4.13.0
urllib3==2.4.0
websockets==15.0.1
Werkzeug==3.1.3
